# app/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Souk API",
        "description": "API du marketplace",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur local"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/categories/create": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Créer une catégorie",
                "description": "Permet de créer une nouvelle catégorie.",
                "operationId": "5106533bd9798ac7c6c66854eaee0541",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "Name"
                                ],
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Électronique"
                                    },
                                    "ParentCategoryID": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "IconURL": {
                                        "description": "Image de l'icône (jpg, jpeg, png, webp - max 2MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Règle métier violée (parent introuvable, slug déjà pris, ...)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Liste des catégories racines (avec filtres, recherche, pagination)",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "hasProducts",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "isEmpty",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Name",
                                "CreatedAt",
                                "UpdatedAt",
                                "DisplayOrder"
                            ]
                        }
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/api/categories/{id}/children": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Sous-catégories directes (avec filtres, recherche, pagination)",
                "operationId": "38ff041ea8914edba5671467cd02ef13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                }
            }
        },
        "/api/categories/{id}/deactivate-subtree": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Désactiver une catégorie et tout son arbre de sous-catégories",
                "operationId": "e452ad4c934ee61fd19eef08d583066c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}/activate": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Activer une catégorie",
                "operationId": "7b0d7ba83314e7722e6401d1f5df49d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Catégorie introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrer uniquement votre JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}